Create README.md
Gets information from the interface: login, password, exchange server, contact number or email to send notification of opening or closing the transaction (can be turned off and on)
Then, with the available processes, one of which you can see, it processes the data using several implemented indicators and then decides to buy, sell or emergency buy and sell. The next activity is to determine the extent of the loss. The user must enter this information, and finally the recommender system produces suggestions with what it has already learned.